relocation interface
