+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from "./Tick";
 export * from "./Event";
 export * from "./StorageData";
-export * from "./interface";
